+++ src/components/logo/logo.jsx
@@ -1,14 +1,12 @@
 import React, { PropTypes } from 'react';
 import { createStyleSheet } from 'jss-theme-reactor';
-import { Icon } from '../../../src';
+import { Icon } from '../../';
 
-const styleSheet = createStyleSheet('Logo', () => (
-  {
-    logo: {
-      display: 'inline-flex',
-    },
-  }
-));
+const styleSheet = createStyleSheet('Logo', () => ({
+  logo: {
+    display: 'inline-flex',
+  },
+}));
 
 function Logo(props, { styleManager }) {
   const classes = styleManager.render(styleSheet);